pyproject.toml
[build-system]
requires = ["setuptools >= 61.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "robo_manip_baselines"
description = "Software that integrates various imitation learning methods and benchmark task environments to provide baselines for robot manipulation"
readme = "README.md"
authors = [
  {name = "Masaki Murooka", email = "m-murooka@aist.go.jp"},
  {name = "Tomohiro Motoda", email = "tomohiro.motoda@aist.go.jp"},
  {name = "Ryoichi Nakajo", email = "ryoichi-nakajo@aist.go.jp"},
]
maintainers = [
  {name = "Masaki Murooka", email = "m-murooka@aist.go.jp"},
]
requires-python = ">= 3.8"
dependencies = [
  "ipdb",
  "ipython",
  "psutil",
  "pyyaml",
  "tqdm",
  "dataclasses",
  "numpy",
  "h5py",
  "scipy>=1.7.0",
  "matplotlib>=3.3.4",
  "opencv-python",
  "ffmpeg-python",
  "imageio>=2.14.1",
  "videoio==0.3.0",
  "gymnasium==1.0.0",
  "mujoco==3.1.6",
  "pin",
  "pyspacemouse",
  "torch",
  "torchvision",
  "torchcodec",
  "tensorboard",
  "pynput>=1.8.1",
]
dynamic = ["version"]

[project.optional-dependencies]
sarnn = [
  "scikit-learn>=1.3.0",
  "torchinfo",
  "gdown",
]
act = [
  "packaging",
  "pyquaternion",
]
mt-act = [
  "packaging",
  "pyquaternion",
  "lru_cache",
  "click",
  "sentence-transformers",
]
diffusion-policy = [
  "cffi",
  "zarr>=2.12.0",
  "numcodecs",
  "hydra-core",
  "einops",
  "dill",
  "diffusers==0.11.1",
  "robomimic==0.2.0",
  "huggingface_hub==0.21.4",
]
diffusion-policy-3d = [
  "zarr==2.12.0",
  "omegaconf",
  "hydra-core==1.2.0",
  "dill==0.3.5.1",
  "einops==0.4.1",
  "diffusers==0.11.1",
  "open3d",
]
flow-policy = [
  "zarr==2.12.0",
  "omegaconf",
  "hydra-core==1.2.0",
  "dill==0.3.5.1",
  "einops==0.4.1",
  "diffusers==0.11.1",
  "open3d",
]
maniflow-policy = [
  "zarr==2.12.0",
  "omegaconf",
  "hydra-core==1.2.0",
  "dill==0.3.5.1",
  "einops==0.4.1",
  "diffusers==0.11.1",
  "open3d",
  "timm",
  "huggingface_hub<=0.25.0"
]
tacto = [
  "pybullet",
  "omegaconf>=2.0.6",
  "pyrender",
  "hydra_core>=1.0.6",
  "lxml",
  "networkx",
  "pycollada",
  "trimesh",
  "pyopengl@git+https://github.com/mmatl/pyopengl.git@76d1261adee2d3fd99b418e75b0416bb7d2865e6",
]
real-ur5e = [
  "ur_rtde",
  "pyrealsense2",
  "hid",
]
real-xarm7 = [
  "xArm-Python-SDK",
  "pyrealsense2",
  "hid",
]
benchmarks = [
  "ruamel.yaml",
  "python-pptx",
  "imageio-ffmpeg",
]

[tool.setuptools.packages.find]
include = ["robo_manip_baselines", "robo_manip_baselines.*"]
